fix: overwrite of generated files by default ones
Before this PR - if you have 2 inputs in globalConfig.json and one
of them already exists in inputs.conf, ucc-gen will not create second
input in the end, because generated inputs.conf would be overwritten
by copying the package folder.

As a sideeffect of this fix - modular alerts configuration files are
now properly generated.