feat: remove placeholder option from globalConfig (#1256)
**Issue number:** ADDON-71391

## Summary

### Changes

> Please provide a summary of what's being changed

The `placeholder` option has been removed from all the applicable
places. The test files that were having this option, it is removed from
them as well. If the `placeholder` isn't present, it passes and the
schema version would be updated to `0.0.8` as we have changed
`schema.json` as well.

### User experience

> Please describe what the user experience looks like before and after
this change

If the developer is using `placeholder`, their build would execute with
the error pointing to resolution, otherwise it would continue as is it
was working before.

## Checklist

If your change doesn't seem to apply, please leave them unchecked.

* [x] I have performed a self-review of this change
* [x] Changes have been tested
* [x] Changes are documented
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0/)

---------

Co-authored-by: Szymon Oleksy <soleksy@splunk.com>