feat: restructure conf and spec file generation (#1328)
**Issue number:** ADDON-73368

## Summary

### Changes

> As per the new design, we are restructuring the code of generating
`.conf` and `.conf.spec` files that are supported by UCC.
One specific change for `inputs.conf.spec`, we are now writing the
default value and description of a field in the spec file.

### User experience

> One change for `inputs.conf.spec`, we are now writing the description
and default value of a field in the spec file if they are mentioned in
the globalConfig, apart from it there are no changes from user end, all
the files that were generated before will be generated as is with these
changes.

## Checklist

If your change doesn't seem to apply, please leave them unchecked.

* [x] I have performed a self-review of this change
* [x] Changes have been tested
* [ ] Changes are documented
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0/)

---------

Co-authored-by: srv-rr-github-token <94607705+srv-rr-github-token@users.noreply.github.com>