fix(globalConfig): update name and rest root convention (#1335)
**Issue number:** ADDON-74883, #1333 
## Summary

### Changes

> Changed the validations for name and restRoot field according to the
conventions specified in ADDON-74883. Additionally, the add-on will not
be built if the name in `globalConfig` does not match the name in
`app.manifest`.

### User experience

> User has to keep the same add-on name in `globalConfig` and
`app.manifest`, if the name does not match then add-on will not be
built. This is usually never the case unless the `app.manifest or name
in globalConfig is changed manually after initializing the add-on.

## Checklist

If your change doesn't seem to apply, please leave them unchecked.

* [x] I have performed a self-review of this change
* [x] Changes have been tested
* [x] Changes are documented
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0/)

---------

Co-authored-by: srv-rr-github-token <94607705+srv-rr-github-token@users.noreply.github.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: sgoral-splunk <138458044+sgoral-splunk@users.noreply.github.com>
Co-authored-by: Artem Rys <rysartem@gmail.com>