fix: prevent unnecessary updates to the globalConfig file (#1535)
**Issue number:** ADDON-77791

### PR Type

**What kind of change does this PR introduce?**
* [ ] Feature
* [x] Bug Fix
* [ ] Refactoring (no functional or API changes)
* [ ] Documentation Update
* [ ] Maintenance (dependency updates, CI, etc.)

## Summary

### Changes

Prevent unnecessary updates to the globalConfig file. To reproduce it -
create a fresh add-on using `init -> build` commands and you shall see
multiple log lines with "Updated globalConfig schema to version
<version>" which should not be needed as it is a completely fresh add-on
which should already use the latest version of the globalConfig file.

The old logic relies on `_uccVersion` field which is also no longer
present (#1519) in the globalConfig file.

The new logic explicitly declares a list of supported schema versions
(which is not the most elegant solution) and uses "0.0.0" in case
`schemaVersion` is not known (this is possible in add-ons built with
older versions of Add-on Builder).

### User experience

No unnecessary updates to the globalConfig file.

## Checklist

If an item doesn't apply to your changes, leave it unchecked.

* [x] I have performed a self-review of this change according to the
[development
guidelines](https://splunk.github.io/addonfactory-ucc-generator/contributing/#development-guidelines)
* [x] Tests have been added/modified to cover the changes [(testing
doc)](https://splunk.github.io/addonfactory-ucc-generator/contributing/#build-and-test)
* [ ] Changes are documented
* [x] PR title and description follows the [contributing
principles](https://splunk.github.io/addonfactory-ucc-generator/contributing/#pull-requests)