feat: add default values of entities to inputs.conf (#1530)
**Issue number:**
ADDON-13362

### PR Type

**What kind of change does this PR introduce?**
* [x] Feature
* [ ] Bug Fix
* [ ] Refactoring (no functional or API changes)
* [ ] Documentation Update
* [ ] Maintenance (dependency updates, CI, etc.)

## Summary
Default values of an entity have been added to `inputs.conf` to
highlight the default configurations for an input.

### Changes
In accordance with the add-on convention, an inputs.conf file should
specify the default values for an input so that the default values are
always present. Therefore, this code change adds default entity values
to `inputs.conf` as defined in globalConfig.

### User experience
User will now be able to point out the default values of their inputs in
`inputs.conf` itself. If an add-on user configuring add-on from .conf
files only, forgets to provide the default values, these changes would
now come from `default/inputs.conf` and they don't have to wait for
error to surface or the add-on to throw out an error.
There would be no change in UX for users using an add-on from GUI or
REST endpoints.

## Checklist

If an item doesn't apply to your changes, leave it unchecked.

* [x] I have performed a self-review of this change according to the
[development
guidelines](https://splunk.github.io/addonfactory-ucc-generator/contributing/#development-guidelines)
* [x] Tests have been added/modified to cover the changes [(testing
doc)](https://splunk.github.io/addonfactory-ucc-generator/contributing/#build-and-test)
* [ ] Changes are documented
* [x] PR title and description follows the [contributing
principles](https://splunk.github.io/addonfactory-ucc-generator/contributing/#pull-requests)