docs: update command flag detail to quickstart.md (#1897)
Current documentation under, **Build the add** on uses  the command: 

`ucc-gen build --source demo_addon_for_splunk/package`

If copied and pasted from the code block and them executed, this would
throw the error: ERROR: Could not determine a version using Git (maybe
not a Git repository?). Use `--ta-version` to specify the version you
want.

This commit adds the missing `--ta-version` to the code block.

**Issue number:**

### PR Type

**What kind of change does this PR introduce?**
* [ ] Feature
* [ ] Bug Fix
* [ ] Refactoring (no functional or API changes)
* [x] Documentation Update
* [ ] Maintenance (dependency updates, CI, etc.)

## Summary

### Changes

Added `--ta-version` flag

### User experience

If a person following the documentation is copying and pasting the
commands they would get an error after executing the command that is
currently in the docs. This is not a blocker because the error is self
explanatory but this part is missing in the docs.

## Checklist

If an item doesn't apply to your changes, leave it unchecked.

### Review

* [ ] self-review - I have performed a self-review of this change
according to the [development
guidelines](https://splunk.github.io/addonfactory-ucc-generator/contributing/#development-guidelines)
* [ ] Changes are documented. The documentation is understandable,
examples work [(more
info)](https://splunk.github.io/addonfactory-ucc-generator/contributing/#documentation-guidelines)
* [ ] PR title and description follows the [contributing
principles](https://splunk.github.io/addonfactory-ucc-generator/contributing/#pull-requests)
* [ ] meeting - I have scheduled a meeting or recorded a demo to explain
these changes (if there is a video, put a link below and in the ticket)

### Tests

See [the testing
doc](https://splunk.github.io/addonfactory-ucc-generator/contributing/#build-and-test).

* [ ] Unit - tests have been added/modified to cover the changes
* [ ] Smoke - tests have been added/modified to cover the changes
* [ ] UI - tests have been added/modified to cover the changes
* [ ] coverage - I have checked the code coverage of my changes [(see
more)](https://splunk.github.io/addonfactory-ucc-generator/contributing/#checking-the-code-coverage)

**Demo/meeting:**

*Reviewers are encouraged to request meetings or demos if any part of
the change is unclear*